+++ shotmanager/__init__.py
@@ -82,11 +82,12 @@ bl_info = {
     "author": "Ubisoft - Julien Blervaque (aka Werwack), Romain Carriquiry Borchiari",
     "description": "Easily manage shots and cameras in the 3D View and see the resulting edit in real-time",
     "blender": (3, 1, 0),
-    "version": (2, 0, 223),
+    "version": (2, 0, 224),
     "location": "View3D > Shot Mng",
     "doc_url": "https://ubisoft-shotmanager.readthedocs.io",
     "tracker_url": "https://github.com/ubisoft/shotmanager/issues",
     #  "warning": "BETA Version",
+    "warning": "Pre-Release",
     "category": "Ubisoft",
 }