+++ shotmanager/ui/sm_ui.py
@@ -91,18 +91,14 @@ class UAS_PT_ShotManager(Panel):
             else:
                 row.label(text=props.project_name)
 
-        addonHeaderWarning = [
-            addon.bl_info.get("warning", "")
-            for addon in addon_utils.modules()
-            if addon.bl_info["name"] == "Shot Manager"
-        ]
-        if False and len(addonHeaderWarning):
+        addonWarning = utils_shot_manager.getShotManagerWanring()
+        if "" != addonWarning:
             betaRow = row.row()
             betaRow.alert = True
-            if "beta" in addonHeaderWarning[0].lower():
+            if "beta" in addonWarning.lower():
                 betaRow.label(text=" ** BETA **")
             else:
-                betaRow.label(text=f" *** {addonHeaderWarning[0]} ***")
+                betaRow.label(text=f" ** {addonWarning} **")
 
     def draw_header_preset(self, context):
         prefs = config.getShotManagerPrefs()